+++ examples/cobra-cmd-ansibleplaybook/cobra-cmd-ansibleplaybook.go
@@ -5,11 +5,13 @@ package main
 */
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"strings"
 
 	ansibler "github.com/apenella/go-ansible"
+	"github.com/apenella/go-ansible/execute"
 	errors "github.com/apenella/go-common-utils/error"
 	"github.com/spf13/cobra"
 )
@@ -73,12 +75,14 @@ func commandHandler(cmd *cobra.Command, args []string) error {
 		Playbook:          playbook,
 		ConnectionOptions: ansiblePlaybookConnectionOptions,
 		Options:           ansiblePlaybookOptions,
-		ExecPrefix:        "Example cobra-cmd-ansibleplaybook",
+		Exec: execute.NewDefaultExecute(
+			execute.WithPrefix("Example cobra-cmd-ansibleplaybook"),
+		),
 	}
 
 	ansibler.AnsibleForceColor()
 
-	err = playbook.Run()
+	err = playbook.Run(context.TODO())
 	if err != nil {
 		return err
 	}