+++ execute/mockExecute.go
@@ -1,31 +1,29 @@
-package ansibler
+package execute
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"io"
 	"os"
+
+	"github.com/apenella/go-ansible/stdoutcallback"
 )
 
 // MockExecute defines a simple executor for testing purposal
 type MockExecute struct {
-	Write     io.Writer
-	CmdRunDir string
+	Write io.Writer
 }
 
 // Execute takes a command and args and runs it, streaming output to stdout
-func (e *MockExecute) Execute(command string, args []string, prefix string) error {
+func (e *MockExecute) Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...ExecuteOptions) error {
 	if e.Write == nil {
 		e.Write = os.Stdout
 	}
-	if command == "error" {
+	if command[0] == "error" {
 		return errors.New("(MockExecute::Execute) error")
 	}
 
-	fmt.Fprintf(e.Write, "%s %v", command, args)
+	fmt.Fprintf(e.Write, "%v", command)
 	return nil
-}
-
-func (e *MockExecute) SetCmdRunDir(Dir string) {
-	e.CmdRunDir = Dir
 }